Moved dataset for deephit to data
